Made `InstructionDialog` component generic and renamed to `NBDialog`
